backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
django-filter==21.1
Pillow==9.5.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
python-dotenv==1.0.0
djoser==2.1.0
drf-extra-fields==3.5.0
drf-yasg==1.21.7 